Move enum to constants and make it a class
